wip: xr and array_backend for upstream module
